+++ scanner/src/test/java/ca/on/oicr/gsi/runscanner/scanner/processor/AbstractProcessorTest.java
@@ -1,11 +1,13 @@
 package ca.on.oicr.gsi.runscanner.scanner.processor;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
 
 import ca.on.oicr.gsi.runscanner.dto.NotificationDto;
 import ca.on.oicr.gsi.runscanner.dto.OxfordNanoporeNotificationDto;
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.databind.JsonMappingException;
+import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.File;
 import java.io.IOException;
@@ -29,10 +31,24 @@ public abstract class AbstractProcessorTest {
           null); // We delete this because it is going to change during updates to dependencies.
       result.setSequencerFolderPath(
           null); // We delete this because it is going to be different in each environment.
-      result.setMetrics(null); // We delete these because changes in metrics are non-critical.
+
+      if (reference.getMetrics() == null) {
+        assertNull(result.getMetrics());
+      } else {
+        // ensure metrics are equal
+        // Load the metrics into a Jackson JsonNode to test for equality
+        ObjectMapper MapperTest = RunProcessor.createObjectMapper();
+        JsonNode jsonNodeResult = MapperTest.readTree(result.getMetrics());
+        JsonNode jsonNodeReference = MapperTest.readTree(reference.getMetrics());
+        assertEquals(jsonNodeReference, jsonNodeResult);
+      }
+
+      // For only Oxford Nanopore processors
       if (clazz.equals(OxfordNanoporeNotificationDto.class)) {
         ((OxfordNanoporeNotificationDto) result).setProtocolVersion(null);
       }
+
+      // For all processors
       assertEquals(reference, result);
     }
   }